Add Guess the Number game in Hardened JS
This JavaScript file creates a simple game of "Guess the Number" using HardenedJS. In this game, a random number between 1 and 100 is chosen when the game starts. The guessTheNumber function can then be called with a number as the guess. If the guess is too low or too high, it informs the player. If the guess is correct, it congratulates the player. This game showcases the use of HardenedJS to provide a secure and interactive gaming experience.